deal with quantity precision

+++ model/src/main/scala/com/github/ppotseluev/algorate/Order.scala
@@ -5,7 +5,7 @@ import enumeratum._
 
 case class Order(
     asset: TradingAsset,
-    lots: Double,
+    lots: BigDecimal,
     operationType: OperationType,
     details: Order.Details,
     info: Order.Info
